// web/src/shared/api/generated/model/weakObjectiveResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Elmanhg.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export interface WeakObjectiveResult {
  objectiveId: string;
  text: string;
  lessonId: string;
  lessonName: string;
  subjectId: string;
  subjectName: string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  servableCount: number | string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  masteredCount: number | string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  seenCount: number | string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  masteryPercent: number | string;
}